mds/mdstypes: move metareqid_t to separate header
