+++ src/mds/QuiesceDb.h
@@ -20,7 +20,7 @@
 #include <optional>
 #include <ranges>
 
-#include "mds/mdstypes.h"
+#include "metareqid_t.h"
 #include "common/ceph_time.h"
 #include "include/cephfs/types.h" // for mds_gid_t
