+++ src/mds/SessionMap.h
@@ -32,7 +32,8 @@
 #include "include/elist.h"
 #include "include/interval_set.h"
 #include "include/object.h" // for object_t
-#include "mdstypes.h" // for metareqid_t, session_info_t
+#include "mdstypes.h" // for session_info_t
+#include "metareqid_t.h"
 #include "mds/MDSAuthCaps.h"
 #include "common/ceph_time.h" // for ceph::coarse_mono_{clock,time}
 #include "common/DecayCounter.h"
